[refactor] Extract DOI component from display component
