+++ spec/components/works/detail_component_spec.rb
@@ -87,73 +87,6 @@ RSpec.describe Works::DetailComponent, type: :component do
     end
   end
 
-  describe 'DOI' do
-    subject(:details) { rendered.css('*').xpath("./table[caption/text()='Details']").first }
-
-    context 'with a DOI' do
-      let(:work_version) { build_stubbed(:work_version, :deposited, version: 2, work: work) }
-      let(:work) { build_stubbed(:work, doi: '10.25740/bc123df4567') }
-
-      it 'renders the doi_link' do
-        expect(details.css('a[href="https://doi.org/10.25740/bc123df4567"]').to_html)
-          .to include 'https://doi.org/10.25740/bc123df4567'
-      end
-    end
-
-    context 'with a collection set to automatically assign a DOI' do
-      let(:collection) { build_stubbed(:collection, doi_option: 'yes') }
-
-      context 'with reserved PURL' do
-        let(:work_version) { build_stubbed(:work_version, :first_draft, work: work) }
-        let(:work) { build_stubbed(:work, collection: collection, assign_doi: false, druid: 'druid:bc123df4567') }
-
-        it 'renders the doi setting' do
-          expect(details.to_html).to include 'DOI will become available once the work has been deposited.'
-        end
-      end
-
-      context 'when it is a first_draft without a reserved PURL' do
-        let(:work_version) { build_stubbed(:work_version, :first_draft, work: work) }
-        let(:work) { build_stubbed(:work, collection: collection, assign_doi: false) }
-
-        it 'renders the doi setting' do
-          expect(details.to_html).to include 'DOI will become available once the work has been deposited.'
-        end
-      end
-
-      context 'when it is a version_draft without a reserved PURL' do
-        let(:work_version) { build_stubbed(:work_version, :version_draft, work: work) }
-        let(:work) { build_stubbed(:work, collection: collection, assign_doi: false) }
-
-        it 'renders the doi setting' do
-          expect(details.to_html).to include 'DOI will become available once a new version is deposited.'
-        end
-      end
-    end
-
-    context 'with a collection set to depositor-selects' do
-      let(:collection) { build_stubbed(:collection, doi_option: 'depositor-selects') }
-
-      context 'when they choose no and have a reserved purl' do
-        let(:work_version) { build_stubbed(:work_version, :first_draft, work: work) }
-        let(:work) { build_stubbed(:work, collection: collection, assign_doi: false, druid: 'druid:bc123df4567') }
-
-        it 'renders the doi setting' do
-          expect(details.to_html).to include 'A DOI has not been assigned to this item.'
-        end
-      end
-
-      context 'when they choose yes without a reserved purl' do
-        let(:work_version) { build_stubbed(:work_version, :first_draft, work: work) }
-        let(:work) { build_stubbed(:work, collection: collection, assign_doi: true) }
-
-        it 'renders the doi setting' do
-          expect(details.to_html).to include 'DOI will become available once the work has been deposited.'
-        end
-      end
-    end
-  end
-
   describe 'DOI settings' do
     context 'with a DOI' do
       let(:work_version) { build_stubbed(:work_version, :deposited, version: 2, work: work) }
